% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewable_residuals.R
\docType{data}
\name{renewable_residuals}
\alias{renewable_residuals}
\title{European Renewable Energy Stocks – Fama–French Residuals (2017–2025)}
\format{
A \code{zoo} object (or data frame) with 2,057 rows and 26 columns.
}
\usage{
renewable_residuals
}
\description{
European Renewable Energy Stocks – Fama–French Residuals (2017–2025)
}
\details{
...
}
\examples{
\dontrun{
data(renewable_residuals)
head(renewable_residuals)
}
}
\keyword{datasets}
